% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_human_footprint}
\alias{load_human_footprint}
\title{Load Human Footprint GeoTIFF}
\usage{
load_human_footprint()
}
\value{
A RasterLayer object containing human footprint data.
}
\description{
This function loads the human footprint GeoTIFF stored in the package's inst/extdata directory.
}
