% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_roads}
\alias{load_roads}
\title{Load Brazilian Roads Shapefile}
\usage{
load_roads()
}
\value{
An sf object containing roads data.
}
\description{
This function loads the roads shapefile stored in the package's inst/extdata directory.
}
